fix(styles): Adjust gear steam theme colors and opacity
Updates the `steam` theme colors for gears to be darker and more muted. Also, increases the fill opacity for gears using the `steam` theme to 0.9 in both the `GearComponent` and `PreviewGear` components for better visibility. Adjusts the `GearProperties` panel to be openable/closable on mobile and fixes a scrollbar issue with the `PreviewGear` component.